Movie_Recommend_System/backend/requirements.txt
flask
flask-cors
pandas
openai
python-dotenv
difflib
